+++ src/components/game/modals/NoteEditModal.tsx
@@ -88,6 +88,8 @@ export function NoteEditModal({
         </div>
       </div>
     </div>,
-    document.body
+    (typeof document !== "undefined" &&
+      document.getElementById("scale-layout-modal-root")) ||
+      document.body
   );
 }